pyproject.toml
[tool.poetry]
name = "vault-py"
version = "0.1.0"
description = ""
authors = ["william <william@twroa.com>"]
readme = "README.md"
packages = [{include = "vault_py"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "2.31.0"
kubernetes = "27.2.0"
jmespath = "1.0.1"
python-dotenv = "1.0.0"
fastapi = "0.108.0"
uvicorn = "0.25.0"
pyjwt = "2.8.0"


[tool.poetry.group.dev.dependencies]
python-dotenv = "1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
